Remove unnecessary cast
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325300 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 3d42b9b430d0403b1092a930190be888f2ce256a

+++ src/jorphan/org/apache/jorphan/collections/Data.java
@@ -226,7 +226,7 @@ public class Data implements Serializable
     {
         int x = start, y = end - 1;
         String basis =
-            ((List) data.get(column)).get((int) ((x + y) / 2)).toString();
+            ((List) data.get(column)).get((x + y) / 2).toString();
         if (x == y)
         {
             return;
